+++ app/features/admin/pagemodal/categorypopover/index.ts
@@ -1,0 +1,1 @@
+export * from "./CategoryPopover";
