+++ app/features/admin/pagemodal/index.ts
@@ -1,0 +1,1 @@
+export * from "./PageModal";
