+++ app/features/admin/pagetable/index.ts
@@ -1,0 +1,2 @@
+export * from "./FilterInput";
+export * from "./PageTable";
