+++ app/features/admin/statistics/index.ts
@@ -1,0 +1,1 @@
+export * from "./Statistics";
