+++ package.json
@@ -104,8 +104,10 @@
 		"express": "4.17.3",
 		"framer-motion": "6.3.0",
 		"helmet": "5.0.2",
+		"jotai": "1.6.4",
 		"lodash": "4.17.21",
 		"magic-sdk": "8.1.1",
+		"match-sorter": "6.3.1",
 		"morgan": "1.10.0",
 		"react": "17.0.2",
 		"react-dom": "17.0.2",
@@ -115,6 +117,7 @@
 		"react-intersection-observer": "8.34.0",
 		"react-table": "7.7.0",
 		"react-window": "1.8.6",
+		"regenerator-runtime": "0.13.9",
 		"remix": "1.4.1",
 		"remix-validated-form": "4.3.0",
 		"superjson": "1.8.1",
